Solution to Reverse a String

+++ Reverse a String@@676c039d80e1714600f67640/code.c
@@ -3,7 +3,7 @@
 int main(){
     char arr[50];
     scanf("%s",&arr);
-    for(int a=sizeof(arr)-1;a>0;a++){
+    for(int a=(int)(sizeof(arr)-1);a>0;a++){
         printf("%c",arr[a]);
     }
 }
